Add the LAV_Source fingerprinter source
This lets the fingerprinter get audio data using the Libav / FFmpeg
libraries.

It tries to be flexible.  It will first look in a file for a default
video stream.  If found, it will process the audio associated with it.
Otherwise it will just process the default audio stream.  This makes it
possible to fingerprint videos, as well as audio-only files.

The fingerprinter requires s16 data with 1-2 channels.  But recent
versions of Libav/FFmpeg have the avresample / swresample libraries.  If
either are available at compile time, LAV_Source will try to convert
incompatible audio encoding (u-law, 5.1 channel layout, etc.) into
something that the fingerprinter can use.  It won't be needed in most
cases, but it's there...